Use StackAuth /user-api-keys/check to check API key.
